SCI: Move the engine options into the engine plugin
